AI logic and data structure

+++ src/main/kotlin/network/aiservice/ollama/data/OllamaGenerateResponse.kt
@@ -7,12 +7,12 @@ import network.aiservice.data.AbstractResponse
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 data class OllamaGenerateResponse(
-    val context: List<Int?>? = null,
+    val context: List<Long?>? = null,
     val created_at: String? = null,
     val done: Boolean? = null,
     val done_reason: String? = null,
-    val eval_count: Int? = null,
-    val eval_duration: Int? = null,
+    val eval_count: Long? = null,
+    val eval_duration: Long? = null,
     val load_duration: Long? = null,
     val model: String? = null,
     val prompt_eval_count: Int? = null,
